requirements.txt
sqlalchemy
asyncpg
psycopg2-binary
alembic
python-dotenv
fastapi
uvicorn
cloudinary
requests
